front-end/src/app/shared/desplegable-permiso/desplegable-permiso.service.spec.ts
import { TestBed } from '@angular/core/testing';

import { DesplegablePermisoService } from './desplegable-permiso.service';
import { TestModule } from '../../core/mocks/test.module';
import { ApiService } from '../../core/services/restClient/api.service';
import { configureTestSuite } from '../../core/mocks/testing';
import { of } from 'rxjs/internal/observable/of';
import { PermisoPadreHijoDataView } from '../data-models/permiso-data-view';

describe('DesplegablePermisoService', () => {
  let apiService: any;
  let service: DesplegablePermisoService;

  configureTestSuite(() => {
    apiService = jasmine.createSpyObj('ApiService', ['get']);
    TestBed.configureTestingModule({
      imports: [TestModule],
      providers: [DesplegablePermisoService,
      {provide: ApiService, useValue: apiService}]
    });
    service = TestBed.get( DesplegablePermisoService);
  });

  beforeEach(() => {
    apiService.get.calls.reset();
  });

  it('should be created', () => {
    expect(service).toBeTruthy();
  });

  describe('El metodo getAll()', () => {
    it('invoca al metodo get del ApiService', () => {
      service.getAll();

      expect(apiService.get).toHaveBeenCalledWith('permisos/padre-hijo');
      expect(apiService.get).toHaveBeenCalledTimes(1);
    });

    it('devuelve un observable con los datos obtenidos por el apiservice.get', () => {
      const esperado = of([{} as PermisoPadreHijoDataView]);
      apiService.get.and.returnValue(esperado);

      const resultado = service.getAll();

      expect(resultado).toEqual(esperado);
    });
  });
});
